Follow flask guidelines
