+++ apps/hash-agents/.gitignore
@@ -1,3 +1,3 @@
 .env
-agents/*/io_types.py
+app/agents/*/io_types.py
 src/agents.ts
