+++ apps/hash-agents/app/agents/template/__main__.py
@@ -11,13 +11,15 @@ def main(agent_input: Input) -> Output:
     return Output(result=float(result[8:]))
 
 
-if __name__ == 'HASH':
+if __name__ == "HASH":
     global IN, OUT
     OUT = main(IN)
 
-if __name__ == '__main__':
-    from .. import setup, get_logger
+if __name__ == "__main__":
+    from ... import setup
+    from logging import getLogger
+
     setup()
 
     output = main(Input(expression="round(pi * 13.37)"))
-    get_logger().info(f"output: {output.result}")
+    getLogger().info(f"output: {output.result}")
