+++ apps/hash-agents/app/agents/template/io_types.ts
